+++ client/scripts/WebSocket.js
@@ -4,9 +4,10 @@ const { MESSAGE_TYPES } = require('../../server/constants')
 export default class WebSocket {
   connection = null
 
-  constructor(url, { onConnection, onLogUpdate } = {}) {
+  constructor(url, { onConnection, onLogUpdate, onAppListUpdate } = {}) {
     this.onConnection = onConnection || _.noop
     this.onLogUpdate = onLogUpdate || _.noop
+    this.onAppListUpdate = onAppListUpdate || _.noop
 
     this.connection = new window.WebSocket(url)
     this.connection.onmessage = this.handleMessageReceive
@@ -20,6 +21,8 @@ export default class WebSocket {
         return this.onConnection(message)
       case MESSAGE_TYPES.LOG:
         return this.onLogUpdate(message)
+      case MESSAGE_TYPES.APPS_LIST_UPDATE:
+        return this.onAppListUpdate(message)
       default:
     }
   }