+++ server/constants.js
@@ -7,7 +7,6 @@ const MESSAGE_TYPES = {
 const PATHS = {
   // paths will been resolved via path.resolve(__dirname)
   WEB_APP_ROOT: '../dist',
-  WEB_APP_HTML: '../dist/index.html',
 }
 
 const TASK_RUNNERS = {